doorbell/app/requirements.txt
flask
flask-cors
Flask
pydub
requests
simpleaudio-patched
